Initial PR fixes
