--- flask_monitoringdashboard/core/alert/README.md
@@ -1,13 +1,0 @@
-# Alerts
-
-## Issue Creation
-
-At the moment the secrets and variables for the issue are saved in a .env file
-
-There is a template with file name `.env_template` that can be used to fill in the info needed for the issue creation alert to work.
-
-
-
-
-
-
